Working on splitting this up.
